Big restructure: split yachtzee into files
